Error Threshold Management (#1555)
* add err_threshold command option


* add err_threshold command option


* config err-threshold


* config err-threshold


* config err-threshold


* config err-threshold


* config err-threshold


* config err-threshold


* config err-threshold


* config err-threshold


* config err-threshold


* config err-threshold


* config err-threshold


* config err-threshold


---------

+++ dozer-core/src/error_manager.rs
@@ -1,4 +1,4 @@
-use std::sync::atomic::AtomicU64;
+use std::sync::atomic::AtomicU32;
 
 use dozer_types::{errors::internal::BoxedError, log::error};
 
@@ -7,22 +7,22 @@ use dozer_types::{errors::internal::BoxedError, log::error};
 /// It panics when an error threshold is set and reached.
 #[derive(Debug)]
 pub struct ErrorManager {
-    threshold: Option<u64>,
-    count: AtomicU64,
+    threshold: Option<u32>,
+    count: AtomicU32,
 }
 
 impl ErrorManager {
-    pub fn new_threshold(threshold: u64) -> Self {
+    pub fn new_threshold(threshold: u32) -> Self {
         Self {
             threshold: Some(threshold),
-            count: AtomicU64::new(0),
+            count: AtomicU32::new(0),
         }
     }
 
     pub fn new_unlimited() -> Self {
         Self {
             threshold: None,
-            count: AtomicU64::new(0),
+            count: AtomicU32::new(0),
         }
     }
